+++ packages/at_onboarding_cli/lib/src/cli/auth_cli.dart
@@ -11,10 +11,13 @@ import 'package:at_lookup/at_lookup.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 import 'package:at_onboarding_cli/src/util/create_at_client_cli.dart';
 import 'package:at_onboarding_cli/src/util/print_full_parser_usage.dart';
+import 'package:at_utils/at_progress.dart';
 import 'package:at_utils/at_utils.dart';
+import 'package:chalkdart/chalk.dart';
 import 'package:duration/duration.dart';
 import 'package:meta/meta.dart';
 
+import '../util/onboarding_util.dart';
 import 'auth_cli_arg_validation.dart';
 import 'auth_cli_args.dart';
 
@@ -72,7 +75,15 @@ Future<int> wrappedMain(List<String> arguments) async {
     arguments = ['onboard', ...arguments];
   }
 
-  final ArgResults topLevelResults = aca.parser.parse(arguments);
+  final ArgResults topLevelResults;
+  try {
+    topLevelResults = aca.parser.parse(arguments);
+  } catch (e) {
+    stderr.writeln('\n$e');
+    stderr.writeln();
+    aca.parser.printAllCommandsUsage();
+    return 1;
+  }
 
   if (topLevelResults.wasParsed(AuthCliArgs.argNameHelp)) {
     aca.sharedArgsParser
@@ -111,7 +122,7 @@ Future<int> wrappedMain(List<String> arguments) async {
   }
 
   // Parse the log levels and act accordingly
-  AtSignLogger.root_level = 'warning';
+  AtSignLogger.root_level = 'shout';
 
   if (commandArgResults.wasParsed(AuthCliArgs.argNameVerbose)) {
     AtSignLogger.root_level = 'info';
@@ -282,13 +293,15 @@ Future<int> wrappedMain(List<String> arguments) async {
     );
     aca.sharedArgsParser.printAllCommandsUsage();
     return 1;
-  } catch (e, st) {
-    stderr.writeln('Error for command ${cliCommand.name}: $e');
-    stderr.writeln(st);
-    commandParser.printAllCommandsUsage(
-      header: 'Usage: ${cliCommand.name}',
-    );
-    aca.sharedArgsParser.printAllCommandsUsage();
+  } catch (e) {
+    await Future.delayed(Duration(milliseconds: 10));
+    stderr.writeln();
+    stderr.writeln();
+    final bolded = chalk.bold(chalk.brightRed('ERROR: ${cliCommand.name}'));
+
+    stderr.writeln('$bolded : $e');
+    stderr.writeln();
+    stderr.writeln('Please try again or contact support@atsign.com');
     return 1;
   }
 
@@ -353,7 +366,7 @@ Future<int> status(ArgResults ar) async {
 /// to send an OTP to the user and then use that OTP to obtain the cram
 /// secret from the registrar.
 @visibleForTesting
-Future<void> onboard(ArgResults argResults, {AtOnboardingService? svc}) async {
+Future<bool> onboard(ArgResults argResults, {AtOnboardingService? svc}) async {
   svc ??= createOnboardingService(argResults);
   logger
       .info('Root server is ${argResults[AuthCliArgs.argNameAtDirectoryFqdn]}');
@@ -361,22 +374,24 @@ Future<void> onboard(ArgResults argResults, {AtOnboardingService? svc}) async {
       'Registrar url provided is ${argResults[AuthCliArgs.argNameRegistrarFqdn]}');
 
   stderr.writeln(
-      '[Information] Onboarding your atSign. This may take up to 2 minutes.');
+      '${chalk.blue('[Information]')} Onboarding your atSign. This may take up to 2 minutes.');
   try {
-    await svc.onboard();
-    return;
-  } on InvalidDataException catch (e) {
-    throw AtEnrollmentException(
-        'Onboarding failed. Please try again. Cause: ${e.message}');
-  } on InvalidRequestException catch (e) {
-    throw AtEnrollmentException(
-        'Onboarding failed. Please try again. Cause: ${e.message}');
-  } on AtActivateException {
-    rethrow;
+    if (argResults[AuthCliArgs.argNameAllowBadRegistrarCerts]) {
+      logger.shout('*************');
+      logger.shout('************* Will ignore bad (expired, invalid, etc)'
+          ' registrar certificates');
+      logger.shout('*************');
+      OnboardingUtil.allowBadCertificates = true;
+    }
+    await svc.onboard(
+      maxRetries: int.parse(argResults[AuthCliArgs.argNameMaxRetries]),
+      retryInterval: AtOnboardingService.defaultActivationCheckInterval,
+    );
+    stderr.writeln();
+    return true;
   } catch (e) {
-    throw ('Onboarding failed.'
-        ' It looks like something went wrong on our side.'
-        ' Please try again or contact support@atsign.com\nCause: $e');
+    await Future.delayed(Duration(milliseconds: 10));
+    throw ('Onboarding failed : $e');
   }
 }
 
@@ -393,7 +408,7 @@ String parseServerResponse(String? response) {
 ///     If it does exist, then the enrollment request has been made and we need
 ///         to try to auth, and act appropriately on the atServer response
 @visibleForTesting
-Future<void> enroll(ArgResults argResults, {AtOnboardingService? svc}) async {
+Future<bool> enroll(ArgResults argResults, {AtOnboardingService? svc}) async {
   if (!argResults.wasParsed(AuthCliArgs.argNameAtKeys)) {
     throw ArgumentError('The --${AuthCliArgs.argNameAtKeys} option is'
         ' mandatory for the "enroll" command');
@@ -402,20 +417,11 @@ Future<void> enroll(ArgResults argResults, {AtOnboardingService? svc}) async {
   File f = File(argResults[AuthCliArgs.argNameAtKeys]);
 
   if (f.existsSync()) {
-    stderr.writeln('Error: atKeys file ${f.path} already exists');
-    return;
+    throw StateError('Error: atKeys file ${f.path} already exists');
   }
 
-  // "canCreateFile" attempts to create the directories for the given [file] if they do not exist,
-  // and then tries to open the file in write mode to verify write permissions.
-  //
-  // If the file can be opened for writing, it is immediately closed and deleted.
-  // In [AtOnboardingServiceImpl._generateAtKeysFile] method, there is a check which returns error
-  // if the file already exists. Therefore, delete the file here after checking for write permissions.
-  //
-  // Incase of any exceptions, the error is logged and returned.
-  if (canCreateFile(f) == false) {
-    return;
+  if (!canCreateFile(f)) {
+    throw StateError('Error: Unable to open $f for writing');
   }
 
   svc ??= createOnboardingService(argResults);
@@ -429,46 +435,28 @@ Future<void> enroll(ArgResults argResults, {AtOnboardingService? svc}) async {
     String permission = l[1].replaceAll('"', '').trim();
     namespaces[namespace] = permission;
   }
-  try {
-    stderr.writeln('Submitting enrollment request');
-    // If apkam Keys expiry is not set, then APKAM keys should lives forever.
-    // Therefore set to 0ms (0 milliseconds) and TTL will not be set.
-    String apkamKeysExpiry = argResults[AuthCliArgs.argNameExpiry] ?? '0ms';
-    AtEnrollmentResponse er = await svc.sendEnrollRequest(
-        argResults[AuthCliArgs.argNameAppName],
-        argResults[AuthCliArgs.argNameDeviceName],
-        argResults[AuthCliArgs.argNamePasscode],
-        namespaces,
-        apkamKeysExpiryDuration: parseDuration(apkamKeysExpiry));
-    stdout.writeln('Enrollment ID: ${er.enrollmentId}');
 
-    stderr.writeln('Waiting for approval; will check every 10 seconds');
-    await svc.awaitApproval(
-      er,
-      retryInterval: Duration(seconds: 10),
+  // If apkam Keys expiry is not set, then APKAM keys should lives forever.
+  // Therefore set to 0ms (0 milliseconds) and TTL will not be set.
+  String apkamKeysExpiry = argResults[AuthCliArgs.argNameExpiry] ?? '0ms';
+  AtEnrollmentResponse er = await svc.sendEnrollRequest(
+      argResults[AuthCliArgs.argNameAppName],
+      argResults[AuthCliArgs.argNameDeviceName],
+      argResults[AuthCliArgs.argNamePasscode],
+      namespaces,
+      apkamKeysExpiryDuration: parseDuration(apkamKeysExpiry));
+  stdout.writeln('Enrollment ID: ${er.enrollmentId}');
+
+  stderr.writeln('Waiting for approval; will check every 10 seconds');
+  await svc.awaitApproval(er,
+      retryInterval: AtOnboardingService.defaultApkamRetryInterval,
       logProgress: true,
-    );
+      maxRetries: int.parse(argResults[AuthCliArgs.argNameMaxRetries]));
 
-    stderr.writeln('Creating atKeys file');
-    await svc.createAtKeysFile(er, allowOverwrite: false);
-  } on InvalidDataException catch (e) {
-    stderr.writeln(
-        '[Error] Enrollment failed. Invalid data provided by user. Please try again\nCause: ${e.message}');
-  } on InvalidRequestException catch (e) {
-    stderr.writeln(
-        '[Error] Enrollment failed. Invalid data provided by user. Please try again\nCause: ${e.message}');
-  } on AtActivateException catch (e) {
-    stderr.writeln('[Error] ${e.message}');
-  } on AtEnrollmentException catch (e) {
-    stderr.writeln('[Fatal] ${e.message}');
-  } on AtAuthenticationException catch (e) {
-    stderr.writeln('[Error] ${e.message}');
-  } on Exception catch (e) {
-    stderr.writeln('[Error] $e');
-    stderr.writeln('[Error] Enrollment failed.\n'
-        '  Cause: $e\n'
-        '  Please try again or contact support@atsign.com');
-  }
+  stderr.writeln('Creating atKeys file');
+  await svc.createAtKeysFile(er, allowOverwrite: false);
+
+  return true;
 }
 
 /// Checks if the specified [file] is writable.
@@ -1043,5 +1031,29 @@ AtOnboardingService createOnboardingService(ArgResults ar) {
     ..hashingAlgoType =
         HashingAlgoType.fromString(ar[AuthCliArgs.argNameHashingAlgoType]);
 
-  return AtOnboardingServiceImpl(atSign, atOnboardingPreference);
+  final impl = AtOnboardingServiceImpl(atSign, atOnboardingPreference);
+  String lastProgressEventType = '';
+  int pad = 10;
+  impl.subscribeProgress().listen((pe) {
+    if (pe.group.isNotEmpty && pe.group != lastProgressEventType) {
+      stderr.writeln();
+    }
+    if (pe.group.length > pad) {
+      pad = pe.group.length;
+    }
+    lastProgressEventType = pe.group;
+    String output = '${pe.type.chalkFn(pe.group.padLeft(pad))} : ${pe.msg}'
+        .replaceAll('\n', '\\n')
+        .replaceAll('\t', ' ');
+    int viewableLength = '${pe.group.padLeft(pad)} : ${pe.msg}'
+        .replaceAll('\n', '\\n')
+        .replaceAll('\t', ' ').length;
+    int diff = output.length - viewableLength;
+    if (stdout.hasTerminal && viewableLength > (stdout.terminalColumns - 3)) {
+      output = '${output.substring(0, stdout.terminalColumns - 3 + diff)}...';
+    }
+    stderr.write('\r\x1b[K$output');
+  });
+
+  return impl;
 }