+++ tests/at_onboarding_cli_functional_tests/test/at_onboarding_cli_test.dart
@@ -176,8 +176,7 @@ void main() {
       },
           throwsA(predicate((dynamic e) =>
               e is AtActivateException &&
-              e.message ==
-                  'atsign $atSign is already activated')));
+              e.message == 'atsign $atSign is already activated')));
     });
 
     tearDown(() async {